@@ -1,11 +1,13 @@
-﻿namespace DataFilters.Grammar.Syntax
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
+﻿using System;
+using System.Linq;
+using System.Text;
+using Candoumbe.Types.Strings;
+using DataFilters.Grammar.Parsing;
+using Microsoft.Extensions.Primitives;
 
-    using static Parsing.FilterTokenizer;
+namespace DataFilters.Grammar.Syntax
+{
+    using static FilterTokenizer;
 
 #if !NETSTANDARD1_3
     using Ardalis.GuardClauses;
@@ -16,12 +18,31 @@
     /// </summary>
     public sealed class EndsWithExpression : FilterExpression, IEquatable<EndsWithExpression>
     {
-        /// <summary>
-        /// The value of the expression
-        /// </summary>
-        public string Value { get; }
-
         private readonly Lazy<string> _lazyEscapedParseableString;
+
+        /// <summary>
+        /// A memory optimized representation of the value hold by the current instance.
+        /// </summary>
+        public StringSegmentLinkedList Value { get; }
+
+        /// <summary>
+        /// Builds a new <see cref="ContainsExpression"/> instance which holds the specified <paramref name="value"/>.
+        /// </summary>
+        /// <param name="value">The desired value</param>
+        /// <exception cref="ArgumentNullException"><paramref name="value"/> contains <see langword="null"/>.</exception>
+        /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is <see cref="StringSegment.Empty"/>.</exception>
+        public EndsWithExpression(StringSegment value) : this(new StringSegmentLinkedList(value))
+        {
+            if (!value.HasValue)
+            {
+                throw new ArgumentNullException(nameof(value));
+            }
+
+            if (StringSegment.IsNullOrEmpty(value))
+            {
+                throw new ArgumentOutOfRangeException(nameof(value));
+            }
+        }
 
         /// <summary>
         /// Builds a new <see cref="EndsWithExpression"/> that holds the specified <paramref name="value"/>.
@@ -29,40 +50,20 @@
         /// <param name="value"></param>
         /// <exception cref="ArgumentNullException"><paramref name="value"/> is <c>null</c>.</exception>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/>'s length is <c>0</c>.</exception>
-        public EndsWithExpression(string value)
+        public EndsWithExpression(StringSegmentLinkedList value)
         {
             Value = value switch
             {
-                null => throw new ArgumentNullException(nameof(value)),
-                {Length: 0} => throw new ArgumentOutOfRangeException(nameof(value)),
+                { Count: 0 } => throw new ArgumentOutOfRangeException(nameof(value)),
                 _ => value
             };
 
             _lazyEscapedParseableString = new Lazy<string>(() =>
             {
-                // The length of the final parseable string in worst cases scenario will double (1 backlash + the escaped character)
-                // Also we need an extra position for the final '*' that will be appended in all cases
-                bool requireEscapingCharacters = value.AtLeastOnce(chr => SpecialCharacters.Contains(chr));
-                StringBuilder parseableString;
+                string escapedValue = Value.Replace(chr => SpecialCharacters.Contains(chr), EscapedSpecialCharacters)
+                    .ToStringValue();
 
-                if (requireEscapingCharacters)
-                {
-                    parseableString = new StringBuilder((value.Length * 2) + 1);
-                    foreach (char chr in value)
-                    {
-                        if (SpecialCharacters.Contains(chr))
-                        {
-                            parseableString = parseableString.Append('\\');
-                        }
-                        parseableString = parseableString.Append(chr);
-                    }
-                }
-                else
-                {
-                    parseableString = new StringBuilder(value, value.Length + 1);
-                }
-
-                return parseableString.Insert(0, '*').ToString();
+                return $"*{escapedValue}";
             });
         }
 
@@ -73,9 +74,9 @@
         /// <exception cref="ArgumentNullException"><paramref name="text"/> is <c>null</c>.</exception>
         public EndsWithExpression(TextExpression text)
 #if !NETSTANDARD1_3
-            : this(Guard.Against.Null(text, nameof(text)).OriginalString)
+            : this(new StringSegmentLinkedList( Guard.Against.Null(text, nameof(text)).OriginalString ))
         {
-            _lazyEscapedParseableString = new(() => $"*{text.EscapedParseableString}");
+            _lazyEscapedParseableString = new Lazy<string>(() => $"*{text.EscapedParseableString}");
         }
 #else
         {
@@ -88,7 +89,7 @@
 #endif
 
         ///<inheritdoc/>
-        public bool Equals(EndsWithExpression other) => Value == other?.Value;
+        public bool Equals(EndsWithExpression other) => Value.Equals(other?.Value);
 
         ///<inheritdoc/>
         public override bool Equals(object obj) => Equals(obj as EndsWithExpression);
