@@ -387,7 +387,7 @@
         [MemberData(nameof(SimplifyCases))]
         public void Given_IntervalExpression_Simplify_should_return_expected_expression(IntervalExpression rangeExpression, FilterExpression expected, string reason)
         {
-            outputHelper.WriteLine($"Range expression : {rangeExpression}");
+            outputHelper.WriteLine($"Range expression : {rangeExpression:d}");
 
             // Act
             FilterExpression actual = rangeExpression.Simplify();
